test: cypress tests
