+++ frontend/cypress/e2e/sidoverskridande.cy.ts
@@ -4,10 +4,9 @@ import { setIntercepts } from 'cypress/support/e2e';
 describe('Sidöverskridande', () => {
   it('Set focus to main', () => {
     cy.visit('/');
-    cy.wait('@getCases');
-    cy.contains('h1', 'Att göra');
+    cy.contains('h1', 'Aktuell förbrukning och produktion');
     cy.contains('a', 'Hoppa till innehåll').then(($link) => {
-      cy.wrap($link).focus().click({ force: true }); // trigger key Enter seem not to work
+      cy.wrap($link).focus().click({ force: true });
     });
     cy.focused().should(($el) => {
       expect($el.prop('tagName')).to.equal('MAIN');
@@ -18,13 +17,12 @@ describe('Sidöverskridande', () => {
     // privat
     cy.viewport('iphone-5');
     cy.visit('/');
-    cy.wait('@getCases');
 
     cy.url().should('include', '/privat/oversikt');
 
     cy.get('button[aria-label="Meny"]').should('be.visible').click();
     cy.get('button[aria-label="Stäng meny"]').should('be.visible');
-    cy.get('ul[aria-label="Undersidor"] li').should('have.length', 4);
+    cy.get('ul[aria-label="Undersidor"] li').should('have.length', 6);
 
     // foretag
     setIntercepts(RepresentingMode.BUSINESS);
@@ -34,7 +32,6 @@ describe('Sidöverskridande', () => {
     cy.get('button[aria-label="Meny"]').should('be.visible').click();
     cy.get('button[aria-label="Stäng meny"]').should('be.visible');
 
-    cy.get('select').contains('option', 'Styrbjörns båtar').should('be.visible').and('be.selected');
     cy.get('select').select('Styrbjörns båtar');
     cy.wait('@postRepresenting').its('response.statusCode').should('eq', 200);
   });
